menu fixed, some php things, not valid yet

+++ profil.php
@@ -3,7 +3,9 @@
 session_start();
 
 include "functions.php"; 
-
+if((!isset($_SESSION["user"])) || empty($_SESSION["user"])){
+	header("Location: login.php");
+}
 
 ?>
 
@@ -24,38 +26,39 @@ include "functions.php";
   
   <nav>
 
-      <div id="navbar">
+  <div id="navbar">
 
-          <div class="logo">
+<div class="logo">
 
-              <div>
-                  <img src="img/kiskakas_logo.png" alt="kiskakas_logo" class="logo">  
-              </div>
+    <div>
+        <img src="img/kiskakas_logo.png" alt="kiskakas_logo" class="logo">  
+    </div>
 
-          </div>
+</div>
 
-          <div id="links">
-              <a href="index.php">Főoldal</a>
-              <a href="rolunk.php">Rólunk</a>
-              <a href="etlap.php">Étlap</a>
-              <a href="itallap.php">Itallap</a>
-              <div class="dropdown">
-                  <button class="dropbtn"><b>Galéria</b></button>
-                  <div class="dropdown-content">
-                      <a href="rendezvenyek.php">Rendezvények</a>
-                      <a href="etelek.php">Ételek</a>
-                  </div>
-              </div>
-            <?php if(!isset($_SESSION["user"]) || empty($_SESSION["user"])): ?>
-                <a href="login.php" >Belépés</a>
-            <?php else: ?>
-                <a href="profil.php" class="active-nav">Profil</a>
-                <a href="logout.php">Kijelentkezés</a>
-            <?php endif; ?>	
+<div id="links">
+    <a href="index.php">Főoldal</a>
+    <a href="rolunk.php">Rólunk</a>
+    <a href="etlap.php">Étlap</a>
+    <a href="itallap.php">Itallap</a>
+    <div class="dropdown">
+        <button class="dropbtn"><b>Galéria</b></button>
+        <div class="dropdown-content">
+            <a href="rendezvenyek.php">Rendezvények</a>
+            <a href="etelek.php">Ételek</a>
+        </div>
+    </div>
+  <?php if(!isset($_SESSION["user"]) || empty($_SESSION["user"])): ?>
+      <a href="login.php" >Belépés</a>
+  <?php else: ?>
+      <a href="profil.php" class="active-nav">Profil</a>
+      <a href="felh.php">Felhasználók</a>
+      <a href="logout.php">Kijelentkezés</a>
+  <?php endif; ?>	
 
-          </div>
+</div>
 
-      </div>
+</div>
 
     </nav>
 
@@ -73,50 +76,43 @@ include "functions.php";
     
     <?php
 
-              $prof ="";
+      if (array_key_exists("profkep",$_FILES)) {
+        $target_dir = "profilkep/";
+        $target_file = $target_dir . basename($_FILES["profkep"]["name"]);
+        $uploadOk = 1;
+        $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
+        if(isset($_POST["feltoltes"])) {
+          $check = getimagesize($_FILES["profkep"]["tmp_name"]);
+          if($check !== false) {
+            echo "A fájl egy kép - " . $check["mime"] . ".";
+            $uploadOk = 1;
+          } else {
+            echo "A fájl nem egy kép.";
+            $uploadOk = 0;
+          }
+        }
 
-              if(!empty($_POST["profkep"])){
-                $prof = $_POST["profkep"];
-              } else {
-                $hiba[] = "Add meg a profilképed!";
-              }
-
-      $target_dir = "profilkep/";
-      $target_file = $target_dir . basename($_FILES["profkep"]["name"]);
-      $uploadOk = 1;
-      $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
-      if(isset($_POST["feltoltes"])) {
-        $check = getimagesize($_FILES["profkep"]["tmp_name"]);
-        if($check !== false) {
-          echo "A fájl egy kép - " . $check["mime"] . ".";
-          $uploadOk = 1;
-        } else {
-          echo "A fájl nem egy kép.";
+        if ($_FILES["profkep"]["size"] > 500000) {
+          echo "Túl nagy méretű fájl.";
           $uploadOk = 0;
         }
-      }
 
-      if ($_FILES["profkep"]["size"] > 500000) {
-        echo "Túl nagy méretű fájl.";
-        $uploadOk = 0;
-      }
-
-      if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg") {
-        echo "Csak JPG, JPEG és PNG formátum feltölthető.";
-        $uploadOk = 0;
-      }
+        if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg") {
+          echo "Csak JPG, JPEG és PNG formátum feltölthető.";
+          $uploadOk = 0;
+        }
 
 
-      if ($uploadOk == 0) {
-        echo "A fájl nem lett feltöltve.";
-      } else {
-        if (move_uploaded_file($_FILES["profkep"]["tmp_name"], $target_file)) {
-          echo "A fájl ". htmlspecialchars( basename( $_FILES["profkep"]["name"])). " feltöltve.";
+        if ($uploadOk == 0) {
+          echo "A fájl nem lett feltöltve.";
         } else {
-          echo "Hiba lépett a feltöltés során.";
+          if (move_uploaded_file($_FILES["profkep"]["tmp_name"], $target_file)) {
+            echo "A fájl ". htmlspecialchars( basename( $_FILES["profkep"]["name"])). " feltöltve.";
+          } else {
+            echo "Hiba lépett a feltöltés során.";
+          }
         }
       }
-
 
     ?>
